refactor(IFileController): rename file
