+++ src/controller/MoveFileController.ts
@@ -2,7 +2,7 @@ import * as path from 'path';
 import { Uri, window } from 'vscode';
 import { FileItem } from '../FileItem';
 import { BaseFileController } from './BaseFileController';
-import { IDialogOptions, IExecuteOptions } from './FileController';
+import { IDialogOptions, IExecuteOptions } from './IFileController';
 
 export interface IMoveFileDialogOptions extends IDialogOptions {
     showFullPath?: boolean;
