+++ src/controller/index.ts
@@ -1,5 +1,5 @@
 export { DuplicateFileController } from './DuplicateFileController';
-export { IFileController } from './FileController';
+export { IFileController } from './IFileController';
 export { MoveFileController } from './MoveFileController';
 export { NewFileController } from './NewFileController';
 export { RemoveFileController } from './RemoveFileController';
